+++ ext/PlotlyJSExtensionModule.jl
@@ -2,6 +2,5 @@ module PlotlyJSExtensionModule
     import PlotlyJS
     import PlotlyExtensionsHelper
 
-    # Load the function for PlotlyBase
-    PlotlyExtensionsHelper._plot_func(::Val{:PlotlyJS}) = PlotlyJS.plot
+    __init__() = PlotlyExtensionsHelper.register_plot_func!(:PlotlyJS, PlotlyJS.plot; priority = 10)
 end